requirements.txt
asgiref==3.4.0
configparser==5.0.2
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
